requirements.txt
tensorflow==2.19.0
flask==3.0.0
numpy==1.24.3
pillow==10.0.0 